requirements.txt
semgrep>=1.45.0
PyYAML>=6.0
requests>=2.28.0
pre-commit>=3.5.0
detect-secrets>=1.4.0
